MAJOR changes to the chat client
It works now so that is a plus. It is pulling messages from the database
no matter the scenario. Two things to note: 1) As guest, it checks to
see if the name that is stored in AGORAController is null. If it is,
make the person's name be "Guest" for the purpose of displaying a name
to the box. 2) When in the lobby it pulls based on the map name "null"
so DO NOT change something that makes things store as null.. That would
mess up everything. 

TODO: 

Michelle, make it look better. The size of the input box was
changed because no matter how big, it made the user input in the middle
AND did not wrap text that hit the end so why did it matter if was huge?
Change as you want. I added the hide chat button just to test the
concept. I like the idea, but feel free to put it somewhere else and not
even necessarily as a button (clickable text is fine). 

Ruiqi and Me, we need to plug this into maps and test if it actually
works with that. I am interested to see!

Other features: Feel free to mention other features that you want
added/taken away. Right now the changing color for specific people is
tougher than you might think lol.

Bugz: The scroll bar works but sometimes it does not display everything
but would be in the textField were the scrollbar NOT there... I have no
idea how to fix that. To see for yourself, give it some time and
eventually you will only see "How bout that... it does!" which is the
last message that does not require scrolling. Move the scrollbar around
even a small bit and it will populate the rest of the box.