Added the functionality to chat while on a map.
Similarly, found a bug... I cannot reproduce it but here are the
specifics: I had just entered the map and I typed a message and clicked
send. It briefly appeared in the window and the disappeared. I typed a
second message and clicked send and it appeared in the box permanently.
I typed a third message and it disappeared again! But the fourth message
stayed. As did the rest of the messages. This means that it is NOT
every other message get. Later I tested entering rapid messages to see
if it was some window that would make it appear and also, nothing so
that rules out if a pull and a push occur together it does not work. The
only thing is something with the first few messages. They literal do not
commit to the DB...