Lots of updates...
1. Implemented the back button for both "My Projects" and "All
Projects". This means that you no longer have to log out and back in if
you want to enter another project. The back button was also made to not
appear until the user has actually entered a project.

2. Fixed the bug that caused "All Projects" to open the same project
that the user opens in "My Projects". Now, the two are different lists.

3. Added a new skin that adjusts the scrolling speed on a map. Now,
instead of the super slow scrolling, it goes the normal speed of 50px.
It can be adjusted to be faster or slower if you desire. I did this just
because I knew it was an area of complaint with your classes and I
stumbled upon the solution while researching other things.

4. Fixed the bug that caused the popups to both log into a project and
move a project to a map to show with no background. Also made it so that
when the user clicks on a project, the password field automatically
gives the cursor so the user does not have to first click in the text
field and then type. For some reason, it wont work for the popup to
"move a map to a project".

5. Deleted all of the empty projects that I created while testing, 4
projects that had a blank title, and two projects that were titled
"Automatically created project". Now, the current release no longer
spills the projects over the box.

6. Changed the name of the first tab to be "A World of Arguments"

7. Fixed a lot of other minor bugs