Added the blue right side bar (see below for way more detail)...
1. Blue side bar added with components "Chat", "Category Chain", and
"Hide", among others.

2. Addition screwed up look of site. Reworked UI of the Menu screen and
login screen. The following details say how and what still doesn't work.
Added percentages for widths and heights instead of hard-coded values
from previous UI developer. Now, the screen will update to fit bigger or
smaller values as the screen size changes. Also, works better for
smaller resolution screens. To note, does NOT work for smallest
resolution (640x480) because of the shear number of components. The
screen resizes but the components get too small to see or begin to
overlap.

3. Fixed bugs related to moving components from previous menu to the
right side bar. This excludes chat. Do not know why that is not working
but I am working on getting it fixed.