Over the next two commits, the following points will be realized...
- debug the fact that you are getting to an empty space on a map when
you previously moved down on a very large map.
- add credits for our funding sources on every blue panel, at the
bottom, that is in the categories and on the maps. Since we have the
least space on the panel for maps, we should do it so small that it
still looks good on the map panel, and then we use the same design for
all the other panels. Here is the text: First line: ["logo-gatech] and
on the right [logo philosophy at gatech]. It should look like on top of
http://philosophy.gatech.edu/, but so small that it fits into one line.
Second line: "Funded by [attached logo FIPSE, probably within a white
box], U.S. Department of Education." I upload the logos to
T-Square-Agora/Resources/Software.
- If I hide the blue panel in a map, it remains hidden also in the
category view. But there it does not bother me much, so it would be
better that the blue panel appears always by default and needs to be
switched off.
- the scroll bar on the bottom does not appear when a map fits on a
screen even though the part on the right might be covered by the blue
panel. That is very confusing because users do not know how to access
stuff that is hidden by the panel (it is not ease to see the "Hide
panel" button. We might increase the button text, but the scroll bar
should appear when anything is hidden from view.
- pop up if user moves mouse over AU in a node or Owner of the map in
the panel: First name user; surname; URL, if provided, and a line
saying: "If you click on this button and the owner of this map provided
a URL, you will be directed to this URL in a new tab of your browser.
Please note that the AGORA-net project is not responsible for the
content of web pages provided by users of the system."
- Screen V (including location of the map; option to change location,
pop-ups, delete option)
- Remove option "Publish your map in the World of Arguments" from maps.
- Add buttons "about us, FAQ, Contact us, Help to the maps.
- we need to change of the text that pops up when user moves mouse over
"Create a new argument map." Please implement the following revision:
"Every map will be created in a private project space that carries your
name and to which only you have access. If you want to publish your map
in the World of Arguments,  go to "My maps" and click on the button that
specifies the location of your map."