Made these tweaks to big yellow box, and translation.xml
The text that pops up when hovering over "Privacy Note" in the category
view contains a link at the end, but it looks really bad in the pop up.
If we cannot embed links in the pop up, we should take the entire last
paragraph from the pop up and put it at the end of the Terms of Service
In the pop up it still promises that user infos are only visible for
registered users, but guest can see that as well. Thus, please remove
the phrase in parentheses from the pop up.
The yellow box after selecting scheme pops up just fine, very good.
However, the box Select: another argument OR complete argument is now
just over the enabler. That is not good because user cannot see what
happens when moving over the different ArgSchemes. It should be on the
side of the enabler.