Added stuff for registration
Did the following:

At the end of the entry boxes there is the button: "I have read and
understood the Privacy Policy." This text needs to be changed into: "I
have read and understood the Privacy Policy, and I agree to the
AGORA-net Terms of Service." (in translation.xml the original is called
<ReadRegistrationNote>. The German needs to be changed into: "Ich habe
die Datenschutz-Vereinbarung gelesen und verstanden, und ich stimme der
Nutzungsvereinbarung zu."
This text needs a check box. Registration can only be completed after
checking the box.
Click on the text above leads to the "Terms of Service" text that pops
up currently in Paul_Testing when you click on "Privacy Policy" on one
of the category pages. This text must be here under registration as
well. For necessary changes to this text, see my remarks below under 2B.
This text is not yet translated. I will do that later and send you the
German translation.
After the line "I have read ...." there needs to be another line with a
check box: “I am 13 years or older or have parental permission to use
AGORA-net.”
Again, registration requires checking the box.