penultimate bug fix before release
Still need to edit scripts so that categories/projects with apostrophes
do not cause an issue when loading.

This fixes project admin screens showing up when a user is not an admin,
a couple issues with refresh and back, maps not showing the correct
message whenever the user edits someone else's message in adversarial
mode, titles/users/owners showing up incorrectly in maps, and maps not
showing correctly in my projects.

Edits to child_projects.php are via lokesh, was having trouble
committing.